switched to float

+++ cpp09/ex01/includes/RPN.hpp
@@ -6,7 +6,7 @@
 /*   By: btenzlin <btenzlin@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/03/02 18:17:39 by btenzlin          #+#    #+#             */
-/*   Updated: 2023/03/07 11:34:00 by btenzlin         ###   ########.fr       */
+/*   Updated: 2023/05/12 22:27:53 by btenzlin         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -29,7 +29,7 @@ public:
 
 	RPN					&operator=(const RPN &other);
 
-	int				calculate(void);
+	float				calculate(void);
 
 private:
